review with refactor chi
